requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
google-generativeai
python-multipart
email-validator
